Convert existing tests to BDD.
