--- src/test/extensionUtilities.test.ts
@@ -1,23 +1,0 @@
-/*!
- * Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import * as assert from 'assert'
-import { safeGet } from '../shared/extensionUtilities'
-
-suite('Extension Utilities Tests', function(): void {
-    class Blah {
-        public someProp?: string
-
-        public constructor(someProp?: string) {
-            this.someProp = someProp
-        }
-    }
-
-    test('nullSafeGet can access sub-property', function() {
-        assert.equal(safeGet(new Blah('hello!'), x => x.someProp), 'hello!')
-        assert.equal(safeGet(new Blah(), x => x.someProp), undefined)
-        assert.equal(safeGet(undefined as Blah | undefined, x => x.someProp), undefined)
-    })
-})
